# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ambulance & Emergency Warehouse Management System API — Damascus
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List equipment with optional filters
      parameters:
        - name: condition
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Paginated list of equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentListResponse"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create new equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: itemType
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string }
        - name: to
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Paginated list of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /transactions/in:
    post:
      operationId: createInTransaction
      tags: [transactions]
      summary: Record an inbound transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/out:
    post:
      operationId: createOutTransaction
      tags: [transactions]
      summary: Record an outbound transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
  /transactions/{id}/print:
    get:
      operationId: getTransactionPrint
      tags: [transactions]
      summary: Get full transaction data for printing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Print data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPrint"
...
  /reports/movements:
    get:
      operationId: getMovementsReport
      tags: [reports]
      summary: Movements report with date range filter
      parameters:
        - name: from
          in: query
          schema: { type: string }
        - name: to
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Movements report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    Equipment:
      type: object
      required: [id, name, condition, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        equipmentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: string
          enum: [good, maintenance, broken, consumed, needs_inspection]
        manufactureYear:
          type: ["integer", "null"]
        originCountry:
          type: ["string", "null"]
        currentHolder:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EquipmentInput:
      type: object
      required: [name, condition]
      properties:
        name:
          type: string
        equipmentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: string
        manufactureYear:
          type: ["integer", "null"]
        originCountry:
          type: ["string", "null"]
        currentHolder:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    EquipmentListResponse:
      type: object
      required: [equipment, total, page, limit]
      properties:
        equipment:
          type: array
          items:
            $ref: "#/components/schemas/Equipment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Transaction:
      type: object
      required: [id, type, itemType, documentNumber, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [in, out, init]
        itemType:
          type: string
          enum: [item, equipment]
        itemId:
          type: ["integer", "null"]
        itemName:
          type: ["string", "null"]
        itemUnit:
          type: ["string", "null"]
        equipmentId:
          type: ["integer", "null"]
        equipmentName:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
        recipientName:
          type: ["string", "null"]
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        exitReason:
          type: ["string", "null"]
        documentNumber:
          type: string
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionListResponse:
      type: object
      required: [transactions, total, page, limit]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TransactionPrint:
      type: object
      required: [transaction, organizationName, printedAt]
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        organizationName:
          type: string
        printedAt:
          type: string
    InTransactionInput:
      type: object
      required: [itemType]
      properties:
        itemType:
          type: string
          enum: [item, equipment]
        itemId:
          type: ["integer", "null"]
        equipmentId:
          type: ["integer", "null"]
        quantity:
          type: ["integer", "null"]
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    OutTransactionInput:
      type: object
      required: [itemType]
      properties:
        itemType:
          type: string
          enum: [item, equipment]
        itemId:
          type: ["integer", "null"]
        equipmentId:
          type: ["integer", "null"]
        quantity:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
...